backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
geoalchemy2
python-dotenv
google-genai
shapely
scikit-learn